Replace pnet::packet with dumbo::pdu
Replaced pnet::packet api with dumbo's, now the code uses both
pnet and dumbo as an intermediate stage, the APIs collide when
it comes to opening the communication channel, the MacAddr
is still from pnet, so it's now copied to a [u8] and a new
MacAddr is made using dumbo's MacAddr.
other changes needed to be made due to incompatibility
between the 2 APIs:

- like in IPv4 packet, dumbo API uses the exact number of
bytes while pnet uses the ihl field and compute the header
length accordingly.

- in dumbo::pdu::ethernet, there's no way to get the raw u8 slice
back after forming the packet, which breaks compat with pnet as
pnet provides the user with a mut [u8] to write their packets into

- used constants from dubmo for udp header size for example since
they're now available